remove using namespace std from debug.h

+++ src/main.cpp
@@ -33,10 +33,10 @@ void Traverse(Polygon p) {
 }
 
 int main() {
-    int n;
+    usize n;
     std::cin >> n;
     std::deque<Point> points;
-    for (int i = 0; i < n; i++) {
+    for (usize i = 0; i < n; i++) {
         double x, y;
         std::cin >> x >> y;
         Point p;
